[Util,Test] Enable & fix tests; fix assertion macro
 - Remove the DISABLED_ prefix from tests for LRUReplacer, BufferPoolManager
 - Correct error in LRUReplacerTest, which unpins frame 1 twice
 - Add extra assertions/expectations to BufferPoolManagerTest

As for the assertion macro: it looks like some parts of the code relied
on it to treat the rest of the function as unreachable if it failed. Of
course, the assert() macro is disabled by default when NDEBUG is defined
(-DNDEBUG), so the linter would complain. I have renamed the existing
macro, and created a new macro with the old name to ensure an exception
is thrown if the condition fails.