[BufferPool,Monolith] Implement BufferPoolManager
 - Remove redundant imports already in header
 - Implement all required methods, using `auto` and private/protected/
   friend methods where convenient

I do lament how similar the implementations for NewPage() and
FetchPage() are. In particular, NewPage() is a subset of FetchPage(),
aside from the distinction between retrieving a particular page ID and
getting a new page ID with AllocatePage(). However, I couldn't seem to
get these to match for the sake of refactoring, and I didn't want to
make the code messier in service of DRY. So instead, all I pulled out
was the new method GetUnpinnedPage().

The only question left unanswered here regards FlushPage(). Does it pin
the page, unpin it, or leave it alone? I've left a note to clarify this
down the line, but for now, I leave it alone.