[Monolith] @ciuji: Add code for Project 3 (Part II)
Set up code for query planning & execution

Deletes a whole bunch of files, while adding some new ones with no
encapsulation (i.e., containing implementations within header files)

If it is later found that some of these files (i.e., the ones that had
not been touched by the student) were not intended to be deleted by the
instructor either--that is, the instructor was simply staging their
work atop an out-of-date copy of the repository--then parts of this
commit may be selectively rolled back. This may be discovered, for
instance, by using Git Blame to discover parts of this commit that
effectively revert the relevant file(s) to a previous state.